Strafe bar fully working
info bar for strafe jumping is finished, in the future i want to add guides to show the optimal angle for speed but that might be hard. godot is still having a persistant issue with loading a scene that isnt the main scene when running the program, but I dont know how to fix that. Merging with main because I think this version is stable